SCRD-2760 Cinder-SES integration
If SES is enabled for Cinder, the SES extension will set the following
global facts:

  - ses_cinder_user_name
  - ses_cinder_pool_name
  - ses_conf_file_path
  - ses_cinder_secret_uuid

We'll then use those facts to configure Cinder accordingly.

NOTE: according to the doc, Cinder can support multiple Ceph storeage pools.
For this iteration, we are assuming one. In the future, we may want to
reconsider this design decision.

Likewise, if SES is enabled for Cinder Backup, the SES extension will set
the following global facts:

  - ses_cinder_user_name
  - ses_cinder_pool_name
  - ses_conf_file_path

We'll update the Cinder Backup configuration accordingly.

Change-Id: Iaa95baacf80321037d0ec011cb52f39e9ba00cf6