Start cinder-volume only if backends configured (bsc#1028382)
This patch tries to test if the rendered cinder.conf for cinder
volume service has a valid enabled_backends setting.  If it doesn't,
then we shouldn't try to start the cinder volume service.

Fix the CND-VOL install & stop tasks to use the platform specific
fact identifying the name of the tgt service.

Change-Id: I6ed148d7fc4c5eb0534d09b4a39c910838cd93f8