filter out API health checks from logs (bsc#1049737)
Prevent the regular health checks performed by haproxy
and the monasca-agent httpcheck plugins from flooding
the log files by employing two magic tricks:

1. use the healthcheck middleware provided by oslo.middleware
with a dedicated /healthcheck URL path. This will no longer
involve any of the other wsgi paste composites, apps and
pipelines when the health check API is pinged, which reduces
the number of entries being logged. When the middleware
is enabled and the service is running, an "OK" message is
returned when the /healthcheck URL path is accessed, which
can be checked by the monasca httpcheck plugin.

2. change the logging configuration to filter out DEBUG and
INFO log entries from eventlet.wsgi.server. These are the only
remaining events being logged after step 1. is enforced.

The net effect is that the regular API calls triggered by the
health checks will not leave any trace in the log files.

Change-Id: Id19c4c79f00f63a28e7c4ea30edfacb13ffca804