Use venv rootwrap precedence for cinder-backup (bsc#1125132)
From all the openstack services that are using rootwrap,
the cinder services are the only ones that use an exec_dirs
configuration in which the system paths take precedence over
the venv paths:

exec_dirs=/sbin,/usr/sbin,/bin,/usr/bin,/usr/local/bin,/usr/local/sbin,
/opt/stack/service/cinder-common-20190208T093735Z/venv/bin

Because of this, the privsep-helper binary is run from
/usr/bin/privsep-helper instead of the venv, which causes it
to ignore the libraries installed in the venv. This leads to
cinder-backup failures:

2019-02-11 21:47:59.490 2079 WARNING oslo.privsep.daemon [-]
privsep log: ImportError: No module named logstash

Reversing the precedence fixes the issue.

Change-Id: I797535bd0845007571b5083ca3d5e51a3c5a51fa