Add option for nfs_share configuration (SOC-9005)
In order to use the netapp NFS driver, the nfs_shares_config
variable should point to a file that contains the hostname/ip
and path to the nfs shares. This patch adds a file called nfs_shares
and installs it on the controllers under the 'etc' dir in the
cinder-volume venv. It also creates the symlink so users can
modify the file under ~/openstack/my_cloud/config/cinder.

Depends-On: Ieca4a25357b329bbd069c55c7fc726ba24631548
Change-Id: Ic00e0517e74f4e1205c7ec8ef9185f452f9d4a08