Ensure nfs-client installed for NetApp support (SOC-9005)
Ensure that the nfs-client package is installed on compute nodes and
nodes running the cinder-volume or cinder-backup services. This is
required to permit NFS mounting of NetApp served volumes either to
attach them to compute instances, or to allow creation of boot from
volume instances, or to permit the volumes to be backed up.

Change-Id: Ic864d23cd82ae179bac6252c651f3d315b9bc63a