Install python-swiftclient as cinder-backup dependency (SOC-11364)
As a workaround for SOC-10522, explicitly install python-swiftclient
on CND-BCK nodes as part of setting up the cinder-backup service.

This avoids a race condition where cinder-backup is started prior to
the Monasca Agent dependencies being installed on the node which may
not happen until after the cinder-backup start is attempted.

Change-Id: I3073bd76e3166ba68ba5b9d94ee928586127de4f